Updating "Next" button and removed indentation
Great script! I've used it for my project but I had to change the class of the "next" button to "artdeco-pagination__button--next", could you please confirm this is the correct class? It worked for me but I would like to make sure it works for you as well!

I've also removed indentation in the while loop so as to not cause an indentation error